+++ packages/example-antd-management-fast-framework/src/pages/Entrance/SignIn/index.jsx
@@ -78,6 +78,9 @@ class SignIn extends BaseComponent {
           },
         );
       },
+      failCallback: ({ target }) => {
+        target.setState({ processing: false });
+      },
     });
   };